Introduce libraries for controlling Varnish.
This branch introduces libraries which can be used in recipes for
installing and configuring Varnish. The libraries are tested on CentOS
6.5, 7.0, and Ubuntu 14.04, with apt, yum, and vendor-cache
packages. None of the existing recipes or attributes have been altered,
so consumers of the cookbook will have a choice on whether or not to use
the cookbook recipes, or the libraries. There are examples in the
fixtures directory on how to use the resources in your recipes, and
documentation in the README has been updated.

In addition, a cloud task has been added to the Rakefile to allow
overriding of Test Kitchen variables. A module was written that detects
the version of Varnish that is running, and Serverspec syntax was
updated on the existing varnish3 and varnish4 spec_helpers.