Support more OS/Varnish combinations, more robust Varnish version selection (#161)
* Fixing a typo

* Remove redundant yum-epel recipe from CentOS test suites

* Exclude varnish from EPEL, pass version to apt_package on Debian

This fixes #160, and also achieves a partial fix for #148. True repo
pinning doesn't seem to be as straightforward as this solution.

* Perform service configuration after application is configured

Testing showed that under Varnish 3 on CentOS 7, systemd would fail to
enable the service at startup, and it was fixed by converging the
instance a second time. Shifting this stanza further down the cookbook
fixed this problem.

* Use attributes to ensure full_stack and distro tests use distro varnish

This setting is set in the install_varnish::distro_install recipe, but
the evaluation of this value happens in attributes.rb before the recipe
has a chance to run. This setting is required for the
install_varnish::full_stack recipe

* Enable testing varnish 3/4 on CentOS 7, and 4/4.1 on Ubuntu 1604

* Stop testing unsupported operating systems that do not work

* Don't use specific package version number when doing distro install

* Move default repo action higher up in the attributes file

* Update CHANGELOG and README

* Increment metadata

* Update note about supported versions